Add Constants Variables For Whole Application
Add constant variables where the whole application can have access to
them because there will parts of the application such as the dispatcher
that will be using the same valuable over and over can when calling the
action.

Also, it uses key mirror which creates an object with values equal to
its key meaning that the key is its own value and since it is a
constant the values are always the name of the keys. This is in order
for the code to be easier to read. Since it is in a hash, the key needs
to be specify in order to access the value and thanks to key mirror the
value will be the name of the key. Allowing not having to guess what
the value is.