Add Flux Actions File For Routes
Add a Flux Actions Creator which are library of helper methods that not
only create the action object but passes the action to the dispatcher.

Different actions are id by the type attribute. When all of the stores
recieve the aciton, they typically use this attribute to determine if
and how they should respond to it.

Also both stores and views control themselves; they are not acted upon
by external objects. Actions flow into the stores through the callbacks
they define and register.