Add Flux Store For Route
Add the Flux Store which is a mix between a model and a controller from
a MVC model. Here the action from the user or application is responded
to. Here is also were to broadcast events to the views that this store
is registered to in order for those views to know that there has been a
change and they should retrieve the state to keep this fresh.

Also, in this file the Store register to the Dispatcher using the
Dispatcher's register function because this will allow the dispatch to
know that this store exist and know which actions the store can be use.